FEAT: Further refactor + tests + comms
